fix: rename folder + remove CSS module
