--- src/components/open/CreateSafe/styles.module.css
@@ -1,7 +1,0 @@
-.back {
-  margin: var(--space-4) 0;
-}
-
-.title {
-  padding-bottom: var(--space-2);
-}
