+++ src/pages/demo.tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from 'next'
 import Head from 'next/head'
-import CreateSafe from '@/components/open/CreateSafe'
+import CreateSafe from '@/components/new-safe/CreateSafe'
 
 const Open: NextPage = () => {
   return (
